% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autonomy-package.R
\docType{package}
\name{autonomy-package}
\alias{autonomy}
\alias{autonomy-package}
\title{autonomy: Build A Package From Rmarkdown File}
\description{
Use Rmarkdown First method to build your package. Start your package with documentation. Everything can be set from a Rmarkdown file in your project.
}
\author{
\strong{Maintainer}: Anna Mantsoki \email{anna.mantsoki@finddx.org} (\href{https://orcid.org/0000-0002-1565-9313}{ORCID})

Other contributors:
\itemize{
  \item FIND [copyright holder]
}

}
\keyword{internal}
